+++ rbx_dom_weak/CHANGELOG.md
@@ -1,6 +1,7 @@
 # rbx\_dom\_weak Changelog
 
 ## Unreleased
+* Added `sort_children_by_key` and `sort_children_unstable_by_key` methods to `RbxInstance` reorder children safely
 
 ## 1.2.0 (2019-03-13)
 * `RbxTree` and `RbxInstance` are now clonable without an `unimplemented` panic